+++ examples/size.py
@@ -4,13 +4,8 @@ from imap_tools import MailBox
 with MailBox('imap.my.ru').login('acc', 'pwd', 'INBOX') as mailbox:
     for msg in mailbox.fetch():
         print(msg.date_str, msg.subject)
-        print('-- RFC822.SIZE message size', msg.size)
-        print('-- bytes size', len(bytes(msg.obj)))
-        print('-- str size', len(str(msg.obj)))
+        print('-- RFC822.SIZE message size', msg.size_rfc822)
+        print('-- bytes size', msg.size)  # will returns headers size only when fetch headers_only=True
         for att in msg.attachments:
             print('---- ATT:', att.filename)
-            print('---- bytes size', len(bytes(att.part)))
-            print('---- str size', len(str(att.part)))
-
-# NOTE:
-# len(bytes(msg.obj)) method will returns headers size only when fetch headers_only=True
+            print('-- bytes size', att.size)